% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R
\name{gp_available}
\alias{gp_available}
\title{Check whether Gunplot or Gnuplot is available}
\usage{
gp_available(path = NULL)
}
\arguments{
\item{path}{Optional executable path or command name.}
}
\value{
\code{TRUE} if an executable can be found, otherwise \code{FALSE}.
}
\description{
Check whether Gunplot or Gnuplot is available
}
