% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{gp_styles}
\alias{gp_styles}
\title{Supported Gnuplot Plotting Styles}
\usage{
gp_styles()
}
\value{
A character vector of plotting style names known to gunplotR.
}
\description{
Supported Gnuplot Plotting Styles
}
