% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R
\name{gp_version}
\alias{gp_version}
\title{Get the Gunplot or Gnuplot version}
\usage{
gp_version(path = NULL)
}
\arguments{
\item{path}{Optional executable path or command name.}
}
\value{
A character vector containing the version output.
}
\description{
Get the Gunplot or Gnuplot version
}
