\docType{package}
\name{gunplotR-package}
\alias{gunplotR}
\alias{gunplotR-package}
\title{Plot from R with Gunplot or Gnuplot}
\description{
A lightweight wrapper for calling an external Gunplot or Gnuplot executable
from R. The package can run plotting scripts and provides a simple XY plotting
helper.
}
